+++ shell/plugins/dashboard-store/__tests__/mutations.spec.ts
